ref: merge two pr-lint phases into one
